add System Audio and organize the audio input popup menu
