+++ AudioAnalyzer/AudioApp.swift
@@ -5,14 +5,16 @@ import ScreenCaptureKit
 final class AudioApp {
     static let shared: AudioApp = .init()
     @Published private(set) var apps: [SCRunningApplication] = []
-    @Published private(set) var displays: [SCDisplay] = []
+    @Published private(set) var display: SCDisplay? // for audio capture, display is required but there are no differences for each displays. we always use the first display.
 
     private init() {
+        reloadApps()
+    }
+
+    func reloadApps() {
         SCShareableContent.getExcludingDesktopWindows(true, onScreenWindowsOnly: false) { [weak self] content, error in
-            NSLog("%@", "content = \(content), windows = \(content?.windows)")
-            self?.displays = content?.displays ?? []
             self?.apps = content?.applications ?? []
-            NSLog("%@", "error = \(String(describing: error))")
+            self?.display = content?.displays.first
         }
     }
 }
